Split rootReducers into recordReducer and userReducer
