+++ src/components/App.js
@@ -18,7 +18,7 @@ function App() {
 }
 
 const mapStateToProps = state => ({
-  user: state.user,
+  user: state.userReducer.user,
 });
 
 const mapDispatchToProps = () => ({
