+++ src/components/Chart.js
@@ -88,8 +88,8 @@ Chart.propTypes = {
 };
 
 const mapStateToProps = state => ({
-  records: state.records,
-  user: state.user,
+  records: state.recordReducer.records,
+  user: state.userReducer.user,
 });
 
 const mapDispatchToProps = () => ({
