+++ src/components/ProtectedLogin.js
@@ -17,7 +17,7 @@ ProtectedLogin.propTypes = {
 };
 
 const mapStateToProps = state => ({
-  auth: state.auth,
+  auth: state.userReducer.auth,
 });
 
 export default connect(mapStateToProps, {})(ProtectedLogin);
