+++ src/reducers/index.js
@@ -1,56 +1,10 @@
-import Cookies from 'js-cookie';
-import {
-  FETCH_USERS,
-  NEW_RECORD,
-  LOGIN,
-  LOGOUT,
-  FETCH_RECORDS,
-} from '../actions/types';
+import { combineReducers } from 'redux';
+import userReducer from './userReducer';
+import recordReducer from './recordReducer';
 
-const user = Cookies.get('user');
+const rootReducer = combineReducers({
+  userReducer,
+  recordReducer,
+});
 
-const initialState = {
-  users: [],
-  user: user ? JSON.parse(user) : {},
-  auth: !!user,
-  records: [],
-};
-
-const userReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case FETCH_USERS:
-      return {
-        ...state,
-        users: action.payload,
-      };
-    case FETCH_RECORDS:
-      return {
-        ...state,
-        records: action.payload,
-      };
-    case NEW_RECORD:
-      return {
-        ...state,
-        records: [...state.records, action.payload],
-      };
-    case LOGIN:
-      Cookies.set('user', action.payload);
-      return {
-        ...state,
-        auth: true,
-        user: action.payload,
-        records: action.records,
-      };
-    case LOGOUT:
-      return {
-        ...state,
-        auth: false,
-        user: {},
-        records: [],
-      };
-    default:
-      return state;
-  }
-};
-
-export default userReducer;
+export default rootReducer;
